List agora é uma função global
-- Refatorado o roteador do websocket para funcionar sem um path na conexão e no re-roteamento (presume home)
-- Os listeners globais foram removidos do index do websocket e colocados em um arquivo 'globalListeners'
-- O evento 'list' agora é global e foi movido da balances para a globalListeners; Seu callback foi arrumado para seguir o padrão 'err, response'
