+++ src/websocket/router/balances.ts
@@ -1,6 +1,5 @@
-import { SuportedCurrencies as SC, CurrencyApi } from '../../currencyApi/currencyApi'
+import { SuportedCurrencies as SC } from '../../currencyApi/currencyApi'
 import currencyApi from '../../currencyApi'
-import User from '../../userApi/user'
 
 /** Interface do objeto experado nos requests de withdraw */
 export interface Withdraw {
@@ -9,35 +8,10 @@ export interface Withdraw {
 	amount: string|number
 }
 
-/** Interface do retorno do socket ao receber 'list' */
-export interface List {
-	code: CurrencyApi['currenciesDetailed'][any]['code']
-	name: CurrencyApi['currenciesDetailed'][any]['name']
-	decimals: CurrencyApi['currenciesDetailed'][any]['decimals']
-	accounts: ReturnType<User['getAccounts']>|undefined
-	balance: string|undefined
-}
-
-export default function balance(socket: SocketIO.Socket) {
+export default function balances(socket: SocketIO.Socket) {
 	console.log('balances endpoint')
 	console.log(`the user is ${socket.user ? '' : 'NOT '}authenticated`)
 	if (!socket.user) return
-
-	/**
-	 * Retorna um array com a lista das currencies, as accounts e o saldo de
-	 * cada currency
-	 */
-	socket.on('list', (callback: (list: List[]) => void) => {
-		console.log('requested balance list')
-		const list = currencyApi.currenciesDetailed.map(currency => ({
-			code:     currency.code,
-			name:     currency.name,
-			decimals: currency.decimals,
-			accounts: socket.user?.getAccounts(currency.name),
-			balance:  socket.user?.getBalance(currency.name).toFullString()
-		}))
-		callback(list)
-	})
 
 	/**
 	 * Executa um request de saque de uma moeda; Retorna o opid da transação
